Added a "main" script as an entry-point.
This is actually on its second iteration now. Originally it accepted
credentials from a specified file or explicitly. The file location is
now implicitly located in the XDG_CONFIG_HOME directory much like the
database.

Previous code has been commented out and will be removed at a later
date (the ability to specify auth file on the command line may be
reinstated).