Use a StaticFileHandler
Although a RequestHandler subclass works, it's the wrong tool for the
job of serving a (quasi-) static image. Browsers cache this
aggressively, preventing the image updating when it changes. For this
purpose, a StaticFileHandler subclass allows control over the caching
HTTP headers which instruct a browser to re-pull the image.