Handle special case of an empty database.
A custom exception has been implemented to inform that the database is
empty. This is expected to be encountered on an initial "update" where
the path of the database has been changed (or the installation is
fresh).

There's another change slipped in here: the `quota_vw` database view
sorts by timestamp (descending, for ease of viewing in sqliteman).