Ensure quota_vw SELECT is sorted.
The quota_vw has been ordered for easy visual access in GUIs that
start with row 1 (this may be reverted!). Additionally, there is
nothing to enforce ordering of quota history/months so it's possible
for row numbers not to reflect the natural order.

This fixes an issue where the graph is plotted "backwards".